requirements.txt
flask==2.2.3
nats-py==2.2.0
flask[async]
aiohttp=3.8.4